Only allow to open local terminal for remote URIs if a ~/.gvfs mapping exists
If GVFS has not been compiled with FUSE support, there exists no ~/.gvfs back-mapping of remote URIs.
Nevertheless, we displayed the possibility to open a local terminal in the corresponding ~/.gvfs path.
This patch introduces a checking routine that determines whether an URI can be mapped to a local path
before showing the local terminal options.