Fix path and SSH remote command quoting.
Up to now, escaping of paths, and of the SSH remote command was broken. We
just nested "" and '' quotes, not properly quoting the strings. This patch attempts
to fix it by using nested g_shell_quote() calls. Please notice that the string passed
to the terminal command is never escaped, while the parameters to the command itself
(i.e. ssh -t's parameter, or cd ...) have to be escaped. Discovered when trying to fix
the unrelated GNOME #583120.